feat: created loading screen for the workout details screen.
