+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/workoutDetails/content/WorkoutDetailsError.kt
@@ -1,0 +1,2 @@
+package com.cairosquad.evolvefit.ui.screen.workoutDetails.content
+
