+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/workoutDetails/content/component/Exercises.kt
@@ -5,10 +5,12 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import com.cairosquad.evolvefit.design_system.component.ExerciseCard
 import com.cairosquad.evolvefit.design_system.theme.Theme
@@ -51,6 +53,7 @@ fun Exercises(
                 time = "",
                 model = exercise.images.firstOrNull() ?: "",
                 modifier = Modifier
+                    .clip(RoundedCornerShape(8.dp))
                     .clickable { onExerciseClick(exercise) },
                 measurementContent = {
                     MeasurementRow(exercise.type)
